inverter-emulator/requirements.txt
# inverter-emulator/requirements.
paho-mqtt==1.6.1
pymodbus==2.5.2
flask==1.1.4
flask_httpauth==4.7.0
MarkupSafe==2.0.1

